PYIC-5881: Add Welsh language test
